+++ pybryt/invariants.py
@@ -1,0 +1,3 @@
+"""Alias for ``pybryt.annotations.invariants``"""
+
+from .annotations.invariants import *